deps(SPV-1286): update eslint and fix errors
