+++ package.json
@@ -62,8 +62,8 @@
     "zod": "^3.24.0"
   },
   "devDependencies": {
-    "@eslint/compat": "^1.0.1",
-    "@eslint/js": "^9.3.0",
+    "@eslint/compat": "^1.2.4",
+    "@eslint/js": "^9.16.0",
     "@playwright/test": "^1.49.1",
     "@tanstack/eslint-plugin-query": "^5.35.6",
     "@tanstack/router-devtools": "^1.33.4",
@@ -75,11 +75,11 @@
     "@vitejs/plugin-react-swc": "^3.7.2",
     "@vitest/coverage-v8": "^2.1.8",
     "autoprefixer": "^10.4.20",
-    "eslint": "^9.8.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-react": "^7.35.0",
-    "eslint-plugin-react-hooks": "^4.6.2",
-    "eslint-plugin-react-refresh": "^0.4.6",
+    "eslint": "^9.16.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-react": "^7.37.2",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "eslint-plugin-react-refresh": "^0.4.16",
     "globals": "^15.3.0",
     "postcss": "^8.4.49",
     "prettier": "3.2.5",
