+++ src/components/ui/form/index.ts
@@ -1,2 +1,9 @@
-export * from './form.tsx';
 export * from './use-form-field.tsx';
+export * from './form-item.tsx';
+export * from './form-item-context.tsx';
+export * from './form-label.tsx';
+export * from './form-message.tsx';
+export * from './form-field.tsx';
+export * from './form-field-context.tsx';
+export * from './form-control.tsx';
+export * from './form-description.tsx';
