+++ src/contexts/AuthContext/index.ts
@@ -1,4 +1,5 @@
 export * from './AuthContext.tsx';
+export * from './AuthProvider.tsx';
 export * from './types.ts';
 export * from './useAuth.tsx';
 export * from './useServerUrl.tsx';
