+++ src/contexts/SpvWalletContext/index.ts
@@ -1,2 +1,3 @@
 export * from './SpvWalletContext.tsx';
+export * from './SpvWalletProvider.tsx';
 export * from './useSpvWalletClient.tsx';
