+++ src/contexts/ThemeProviderContext/index.ts
@@ -1,2 +1,3 @@
 export * from './ThemeProviderContext.tsx';
+export * from './ThemeProvider.tsx';
 export * from './useTheme.tsx';
